Returning an error if the socket fails to connect.
The sockets now also have an error event listener that will call the callback if an error occurs before the connection is established. The lack of this listener caused the driver to stop the process because an error was thrown.
If an error occurs after the connection is established, this will only be logged if the logOperations options is set to true.